+++ src/client/js/otp/modules/planner/ItinerariesWidget.js
@@ -474,6 +474,9 @@ otp.widgets.ItinerariesWidget =
             itinDiv.append("<div class='otp-itinSysNoticeRow'><b>System tags</b>: " + systemTags + "</div>");
         }
         itinDiv.append(itinAccord);
+        if (itin.itinData.arrivedAtDestinationWithRentedBicycle) {
+            itinDiv.append("<div class='otp-itinAlertRow'><b>" + _tr("Arrived at destination with a rented bicycle!") + "</b></div>");
+        }
         //TRANSLATORS: End: Time and date (Shown after path itinerary)
         itinDiv.append("<div class='otp-itinEndRow'><b>" + _tr("End") + "</b>: "+itin.getEndTimeStr()+"</div>");