+++ src/main/java/org/opentripplanner/updater/bike_rental/datasources/params/GbfsBikeRentalDataSourceParameters.java
@@ -4,10 +4,12 @@ import org.opentripplanner.updater.DataSourceType;
 
 public class GbfsBikeRentalDataSourceParameters extends BikeRentalDataSourceParameters {
 
+  private final boolean allowKeepingRentedBicycleAtDestination;
   private final boolean routeAsCar;
 
-  public GbfsBikeRentalDataSourceParameters(String url, String network, boolean routeAsCar) {
+  public GbfsBikeRentalDataSourceParameters(String url, String network, boolean routeAsCar, boolean allowKeepingRentedBicycleAtDestination) {
     super(DataSourceType.GBFS, url, network, null);
+    this.allowKeepingRentedBicycleAtDestination = allowKeepingRentedBicycleAtDestination;
     this.routeAsCar = routeAsCar;
   }
 
@@ -18,4 +20,7 @@ public class GbfsBikeRentalDataSourceParameters extends BikeRentalDataSourcePara
   @Override
   public String getApiKey() { throw new UnsupportedOperationException(); }
 
+  public boolean allowKeepingRentedBicycleAtDestination() {
+    return allowKeepingRentedBicycleAtDestination;
+  }
 }