+++ src/test/java/org/opentripplanner/routing/algorithm/BikeRentalTest.java
@@ -14,6 +14,7 @@ import org.opentripplanner.routing.edgetype.StreetEdge;
 import org.opentripplanner.routing.edgetype.StreetTraversalPermission;
 import org.opentripplanner.routing.graph.Graph;
 import org.opentripplanner.routing.graph.Vertex;
+import org.opentripplanner.routing.location.TemporaryStreetLocation;
 import org.opentripplanner.routing.vertextype.BikeRentalStationVertex;
 import org.opentripplanner.routing.vertextype.StreetVertex;
 import org.opentripplanner.routing.vertextype.TransitEntranceVertex;
@@ -28,6 +29,7 @@ public class BikeRentalTest extends GraphRoutingTest {
 
     private Graph graph;
     private TransitStopVertex S1;
+    private TemporaryStreetLocation T1, T2;
     private TransitEntranceVertex E1;
     private StreetVertex A, B, C, D;
     private BikeRentalStationVertex B1, B2;
@@ -42,6 +44,7 @@ public class BikeRentalTest extends GraphRoutingTest {
         //   B <-> B1
         //   C <-> B2
         //   D <-> E1
+        //   D <-> T2
 
         graph = graphOf(new Builder() {
             @Override
@@ -53,6 +56,9 @@ public class BikeRentalTest extends GraphRoutingTest {
                 D = intersection("D", 47.530, 19.000);
                 E1 = entrance("E1", 47.530, 19.001);
 
+                T1 = streetLocation("T1", 47.500, 18.999, false);
+                T2 = streetLocation("T1", 47.530, 18.999, true);
+
                 B1 = bikeRentalStation("B1", 47.510, 19.001);
                 B2 = bikeRentalStation("B2", 47.520, 19.001);
 
@@ -61,6 +67,9 @@ public class BikeRentalTest extends GraphRoutingTest {
 
                 biLink(B, B1);
                 biLink(C, B2);
+
+                link(T1, A);
+                link(D, T2);
 
                 SE1 = street(A, B, 50, StreetTraversalPermission.PEDESTRIAN_AND_BICYCLE);
                 SE2 = street(B, C, 1000, StreetTraversalPermission.PEDESTRIAN_AND_BICYCLE);
@@ -101,7 +110,8 @@ public class BikeRentalTest extends GraphRoutingTest {
                 S1, E1,
                 "WALK - BEFORE_RENTING - AB street (76.19, 38)",
                 "BICYCLE - RENTING_FROM_STATION - BC street (540.19, 280)",
-                "WALK - HAVE_RENTED - CD street (650.38, 333)"
+                "WALK - HAVE_RENTED - CD street (650.38, 333)",
+                "WALK - HAVE_RENTED - E1 (651.38, 333)"
         );
     }
 
@@ -114,12 +124,14 @@ public class BikeRentalTest extends GraphRoutingTest {
                 List.of(
                         "WALK - BEFORE_RENTING - AB street (76.19, 38)",
                         "WALK - BEFORE_RENTING - BC street (1,579.95, 790)",
-                        "WALK - BEFORE_RENTING - CD street (1,655.14, 828)"
+                        "WALK - BEFORE_RENTING - CD street (1,655.14, 828)",
+                        "WALK - BEFORE_RENTING - E1 (1,656.14, 828)"
                 ),
                 List.of(
                         "WALK - HAVE_RENTED - AB street (76.19, 38)",
                         "WALK - HAVE_RENTED - BC street (1,579.95, 790)",
-                        "WALK - HAVE_RENTED - CD street (1,655.14, 828)"
+                        "WALK - HAVE_RENTED - CD street (1,655.14, 828)",
+                        "WALK - HAVE_RENTED - E1 (1,656.14, 828)"
                 )
         );
     }
@@ -133,12 +145,14 @@ public class BikeRentalTest extends GraphRoutingTest {
                 List.of(
                         "WALK - BEFORE_RENTING - AB street (76.19, 38)",
                         "WALK - BEFORE_RENTING - BC street (1,579.95, 790)",
-                        "WALK - BEFORE_RENTING - CD street (1,655.14, 828)"
+                        "WALK - BEFORE_RENTING - CD street (1,655.14, 828)",
+                        "WALK - BEFORE_RENTING - E1 (1,656.14, 828)"
                 ),
                 List.of(
                         "WALK - HAVE_RENTED - AB street (76.19, 38)",
                         "WALK - HAVE_RENTED - BC street (1,579.95, 790)",
-                        "WALK - HAVE_RENTED - CD street (1,655.14, 828)"
+                        "WALK - HAVE_RENTED - CD street (1,655.14, 828)",
+                        "WALK - HAVE_RENTED - E1 (1,656.14, 828)"
                 )
         );
     }
@@ -152,12 +166,14 @@ public class BikeRentalTest extends GraphRoutingTest {
                 List.of(
                         "WALK - BEFORE_RENTING - AB street (76.19, 38)",
                         "WALK - BEFORE_RENTING - BC street (1,579.95, 790)",
-                        "WALK - BEFORE_RENTING - CD street (1,655.14, 828)"
+                        "WALK - BEFORE_RENTING - CD street (1,655.14, 828)",
+                        "WALK - BEFORE_RENTING - E1 (1,656.14, 828)"
                 ),
                 List.of(
                         "WALK - HAVE_RENTED - AB street (76.19, 38)",
                         "WALK - HAVE_RENTED - BC street (1,579.95, 790)",
-                        "WALK - HAVE_RENTED - CD street (1,655.14, 828)"
+                        "WALK - HAVE_RENTED - CD street (1,655.14, 828)",
+                        "WALK - HAVE_RENTED - E1 (1,656.14, 828)"
                 )
         );
     }
@@ -170,7 +186,8 @@ public class BikeRentalTest extends GraphRoutingTest {
                 S1, E1, false,
                 "WALK - BEFORE_RENTING - AB street (76.19, 38)",
                 "BICYCLE - RENTING_FROM_STATION - BC street (540.19, 280)",
-                "WALK - HAVE_RENTED - CD street (650.38, 333)"
+                "WALK - HAVE_RENTED - CD street (650.38, 333)",
+                "WALK - HAVE_RENTED - E1 (651.38, 333)"
         );
     }
 
@@ -182,7 +199,8 @@ public class BikeRentalTest extends GraphRoutingTest {
                 S1, E1, false,
                 "WALK - BEFORE_RENTING - AB street (76.19, 38)",
                 "BICYCLE - RENTING_FROM_STATION - BC street (540.19, 280)",
-                "WALK - HAVE_RENTED - CD street (650.38, 333)"
+                "WALK - HAVE_RENTED - CD street (650.38, 333)",
+                "WALK - HAVE_RENTED - E1 (651.38, 333)"
         );
     }
 
@@ -194,7 +212,136 @@ public class BikeRentalTest extends GraphRoutingTest {
                 S1, E1,
                 "WALK - BEFORE_RENTING - AB street (76.19, 38)",
                 "BICYCLE - RENTING_FLOATING - BC street (540.19, 280)",
-                "BICYCLE - RENTING_FLOATING - CD street (560.19, 290)"
+                "BICYCLE - RENTING_FLOATING - CD street (560.19, 290)",
+                "BICYCLE - RENTING_FLOATING - E1 (561.19, 290)"
+        );
+
+        assertPath(
+                S1, T2,
+                "WALK - BEFORE_RENTING - AB street (76.19, 38)",
+                "BICYCLE - RENTING_FLOATING - BC street (540.19, 280)",
+                "BICYCLE - RENTING_FLOATING - CD street (560.19, 290)",
+                "null - RENTING_FLOATING - null (561.19, 290)"
+        );
+
+        assertPath(
+                T1, E1,
+                "WALK - BEFORE_RENTING - AB street (76.19, 38)",
+                "BICYCLE - RENTING_FLOATING - BC street (540.19, 280)",
+                "BICYCLE - RENTING_FLOATING - CD street (560.19, 290)",
+                "BICYCLE - RENTING_FLOATING - E1 (561.19, 290)"
+        );
+    }
+
+    @Test
+    public void testBikeRentalFromStationWantToKeepCantKeep() {
+        B1.getStation().isKeepingBicycleRentalAtDestinationAllowed = false;
+
+        assertPath(
+                S1, E1, 40,
+                "WALK - BEFORE_RENTING - AB street (76.19, 38)",
+                "BICYCLE - RENTING_FROM_STATION - BC street (540.19, 280)",
+                "WALK - HAVE_RENTED - CD street (650.38, 333)",
+                "WALK - HAVE_RENTED - E1 (651.38, 333)"
+        );
+
+        assertPath(
+                S1, T2, 40,
+                "WALK - BEFORE_RENTING - AB street (76.19, 38)",
+                "BICYCLE - RENTING_FROM_STATION - BC street (540.19, 280)",
+                "WALK - HAVE_RENTED - CD street (650.38, 333)",
+                "null - HAVE_RENTED - null (651.38, 333)"
+        );
+
+        assertPath(
+                T1, E1, 40,
+                "WALK - BEFORE_RENTING - AB street (76.19, 38)",
+                "BICYCLE - RENTING_FROM_STATION - BC street (540.19, 280)",
+                "WALK - HAVE_RENTED - CD street (650.38, 333)",
+                "WALK - HAVE_RENTED - E1 (651.38, 333)"
+        );
+    }
+
+    @Test
+    public void testBikeRentalFromStationWantToKeepCanKeep() {
+        B1.getStation().isKeepingBicycleRentalAtDestinationAllowed = true;
+
+        assertPath(
+                S1, E1, 40,
+                "WALK - BEFORE_RENTING - AB street (76.19, 38)",
+                "BICYCLE - RENTING_FROM_STATION (may keep) - BC street (540.19, 280)",
+                "BICYCLE - RENTING_FROM_STATION (may keep) - CD street (560.19, 290)",
+                "BICYCLE - RENTING_FROM_STATION (may keep) - E1 (601.19, 290)"
+        );
+
+        assertPath(
+                S1, T2, 40,
+                "WALK - BEFORE_RENTING - AB street (76.19, 38)",
+                "BICYCLE - RENTING_FROM_STATION (may keep) - BC street (540.19, 280)",
+                "BICYCLE - RENTING_FROM_STATION (may keep) - CD street (560.19, 290)",
+                "null - RENTING_FROM_STATION (may keep) - null (601.19, 290)"
+        );
+
+        assertPath(
+                T1, E1, 40,
+                "WALK - BEFORE_RENTING - AB street (76.19, 38)",
+                "BICYCLE - RENTING_FROM_STATION (may keep) - BC street (540.19, 280)",
+                "BICYCLE - RENTING_FROM_STATION (may keep) - CD street (560.19, 290)",
+                "BICYCLE - RENTING_FROM_STATION (may keep) - E1 (601.19, 290)"
+        );
+    }
+
+    @Test
+    public void testBikeRentalFromStationWantToKeepCanKeepButCostly() {
+        B1.getStation().isKeepingBicycleRentalAtDestinationAllowed = true;
+        int keepRentedBicycleAtDestinationCost = 1000;
+
+        assertPath(
+                S1, E1, false, keepRentedBicycleAtDestinationCost,
+                List.of(
+                        "WALK - BEFORE_RENTING - AB street (76.19, 38)",
+                        "BICYCLE - RENTING_FROM_STATION (may keep) - BC street (540.19, 280)",
+                        "WALK - HAVE_RENTED - CD street (650.38, 333)",
+                        "WALK - HAVE_RENTED - E1 (651.38, 333)"
+                ),
+                List.of(
+                        "WALK - BEFORE_RENTING - AB street (76.19, 38)",
+                        "BICYCLE - RENTING_FROM_STATION - BC street (540.19, 280)",
+                        "WALK - HAVE_RENTED - CD street (650.38, 333)",
+                        "WALK - HAVE_RENTED - E1 (651.38, 333)"
+                )
+        );
+
+        assertPath(
+                S1, T2, false, keepRentedBicycleAtDestinationCost,
+                List.of(
+                        "WALK - BEFORE_RENTING - AB street (76.19, 38)",
+                        "BICYCLE - RENTING_FROM_STATION (may keep) - BC street (540.19, 280)",
+                        "WALK - HAVE_RENTED - CD street (650.38, 333)",
+                        "null - HAVE_RENTED - null (651.38, 333)"
+                ),
+                List.of(
+                        "WALK - BEFORE_RENTING - AB street (76.19, 38)",
+                        "BICYCLE - RENTING_FROM_STATION - BC street (540.19, 280)",
+                        "WALK - HAVE_RENTED - CD street (650.38, 333)",
+                        "null - HAVE_RENTED - null (651.38, 333)"
+                )
+        );
+
+        assertPath(
+                T1, E1, false, keepRentedBicycleAtDestinationCost,
+                List.of(
+                        "WALK - BEFORE_RENTING - AB street (76.19, 38)",
+                        "BICYCLE - RENTING_FROM_STATION (may keep) - BC street (540.19, 280)",
+                        "WALK - HAVE_RENTED - CD street (650.38, 333)",
+                        "WALK - HAVE_RENTED - E1 (651.38, 333)"
+                ),
+                List.of(
+                        "WALK - BEFORE_RENTING - AB street (76.19, 38)",
+                        "BICYCLE - RENTING_FROM_STATION - BC street (540.19, 280)",
+                        "WALK - HAVE_RENTED - CD street (650.38, 333)",
+                        "WALK - HAVE_RENTED - E1 (651.38, 333)"
+                )
         );
     }
 
@@ -217,15 +364,46 @@ public class BikeRentalTest extends GraphRoutingTest {
     private void assertPath(
             Vertex fromVertex,
             Vertex toVertex,
+            int keepRentedBicycleCost,
+            String... descriptor
+    ) {
+        assertPath(
+                fromVertex, toVertex, false, keepRentedBicycleCost, List.of(descriptor),
+                List.of(descriptor)
+        );
+    }
+
+    private void assertPath(
+            Vertex fromVertex,
+            Vertex toVertex,
             boolean useAvailabilityInformation,
             List<String> departAtDescriptor,
             List<String> arriveByDescriptor
     ) {
         List<String> departAt = runStreetSearchAndCreateDescriptor(fromVertex, toVertex, false,
-                useAvailabilityInformation
+                useAvailabilityInformation, 0
         );
         List<String> arriveBy = runStreetSearchAndCreateDescriptor(fromVertex, toVertex, true,
-                useAvailabilityInformation
+                useAvailabilityInformation, 0
+        );
+
+        assertEquals(departAtDescriptor, departAt, "departAt path");
+        assertEquals(arriveByDescriptor, arriveBy, "arriveBy path");
+    }
+
+    private void assertPath(
+            Vertex fromVertex,
+            Vertex toVertex,
+            boolean useAvailabilityInformation,
+            int keepBicycleRentalCost,
+            List<String> departAtDescriptor,
+            List<String> arriveByDescriptor
+    ) {
+        List<String> departAt = runStreetSearchAndCreateDescriptor(fromVertex, toVertex, false,
+                useAvailabilityInformation, keepBicycleRentalCost
+        );
+        List<String> arriveBy = runStreetSearchAndCreateDescriptor(fromVertex, toVertex, true,
+                useAvailabilityInformation, keepBicycleRentalCost
         );
 
         assertEquals(departAtDescriptor, departAt, "departAt path");
@@ -236,7 +414,8 @@ public class BikeRentalTest extends GraphRoutingTest {
             Vertex fromVertex,
             Vertex toVertex,
             boolean arriveBy,
-            boolean useAvailabilityInformation
+            boolean useAvailabilityInformation,
+            int keepRentedBicycleCost
     ) {
         var options = new RoutingRequest();
         options.arriveBy = arriveBy;
@@ -246,6 +425,8 @@ public class BikeRentalTest extends GraphRoutingTest {
         options.bikeRentalDropoffTime = 15;
         options.useBikeRentalAvailabilityInformation = useAvailabilityInformation;
         options.worstTime = arriveBy ? Long.MIN_VALUE : Long.MAX_VALUE;
+        options.allowKeepingRentedBicycleAtDestination = keepRentedBicycleCost > 0;
+        options.keepingRentedBicycleAtDestinationCost = keepRentedBicycleCost;
 
         return runStreetSearchAndCreateDescriptor(
                 fromVertex, toVertex, arriveBy, options, StreetMode.BIKE_RENTAL);
@@ -273,12 +454,13 @@ public class BikeRentalTest extends GraphRoutingTest {
 
         return path.states
                 .stream()
-                .filter(s -> s.getBackEdge() instanceof StreetEdge)
+                .filter(s -> s.getBackEdge() instanceof StreetEdge || s.getVertex() == toVertex)
                 .map(s -> String.format(
                         Locale.ROOT,
-                        "%s - %s - %s (%,.2f, %d)",
+                        "%s - %s%s - %s (%,.2f, %d)",
                         s.getBackMode(),
                         s.getBikeRentalState(),
+                        s.mayKeepRentedBicycleAtDestination() ? " (may keep)" : "",
                         s.getBackEdge() != null ? s.getBackEdge().getName() : null,
                         s.getWeight(),
                         s.getElapsedTimeSeconds()